+++ util/str/str.go
@@ -6,16 +6,16 @@ import "strings"
 // @Author:  Joshua Conero
 // @Name:    字符串
 
-func Ucfirst(str string) string  {
+func Ucfirst(str string) string {
 	idx := strings.Index(str, " ")
-	if idx > -1{
+	if idx > -1 {
 		newStr := []string{}
-		for _, s := range strings.Split(str, " "){
+		for _, s := range strings.Split(str, " ") {
 			newStr = append(newStr, Ucfirst(s))
 		}
 		str = strings.Join(newStr, "")
-	}else{
-		if len(str) > 0{
+	} else {
+		if len(str) > 0 {
 			str = strings.ToUpper(str[0:1]) + str[1:]
 		}
 	}